requirements.txt
requests
dnspython
python-whois
ipwhois
ipaddress
python-nmap
